V0.3.10 release
V0.3.10 release
The project started with only a few files so the project folder structure was flat.
This release refactors the overall project organization into a more standardized folder structure.
		Project
			/docs
			/forms
			/include
			/src
			CmakeList.txt
Also added is the ability to read the current workmode from the L2 and get tha L2 Parameters packet.

+++ src/AxisGridRenderer.cpp
@@ -114,7 +114,7 @@ void AxisGridRenderer::initialize()
 
     m_vbo.create();
     m_vbo.bind();
-    m_vbo.allocate(vertices.constData(), vertices.size() * sizeof(Vertex));
+    m_vbo.allocate(vertices.constData(), (int)(vertices.size() * sizeof(Vertex)));
 
     glEnableVertexAttribArray(0); // position
     glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(Vertex), nullptr);
